feat:procurement menambah alert

+++ resources/views/pages/data-master/procurement/_function/function.blade.php
@@ -27,17 +27,36 @@
 
     function deleteFunc(id) {
         var id = id;
-        // ajax
-        $.ajax({
-            type: "DELETE",
-            url: "{{ route('procurement.destroy') }}",
-            data: {
-                id: id
-            },
-            dataType: 'json',
-            success: function(res) {
-                var oTable = $('#myTable').dataTable();
-                oTable.fnDraw(false);
+        Swal.fire({
+            title: 'Apakah Anda yakin?',
+            text: "Data yang dihapus tidak dapat dikembalikan!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#d33',
+            cancelButtonColor: '#3085d6',
+            confirmButtonText: 'Ya, Hapus!',
+            cancelButtonText: 'Batal'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                $.ajax({
+                    type: "DELETE",
+                    url: "{{ route('procurement.destroy') }}",
+                    data: {
+                        id: id
+                    },
+                    dataType: 'json',
+                    success: function(res) {
+                        var oTable = $('#myTable').dataTable();
+                        oTable.fnDraw(false);
+                        Swal.fire({
+                            icon: 'success',
+                            title: 'Success',
+                            text: res.message,
+                            timer: 2000,
+                            showConfirmButton: false
+                        })
+                    }
+                });
             }
         });
     }
@@ -52,12 +71,32 @@
             cache: false,
             contentType: false,
             processData: false,
-            success: (data) => {
+            success: (response) => {
+                console.log('====================================');
+                console.log(response);
+                console.log('====================================');
                 $("#procurement-modal").modal('hide');
                 var oTable = $('#myTable').dataTable();
                 oTable.fnDraw(false);
                 $("#btn-save").html('Submit');
                 $("#btn-save").attr("disabled", false);
+                if (response.success == true) {
+                    Swal.fire({
+                        icon: 'success',
+                        title: 'Success',
+                        text: response.message,
+                        timer: 1000,
+                        showConfirmButton: false
+                    });
+                } else if (response.success == false) {
+                    Swal.fire({
+                        icon: 'error',
+                        title: 'Failed',
+                        text: response.message,
+                        timer: 2000,
+                        showConfirmButton: false
+                    });
+                }
             },
             error: function(data) {
                 console.log(data);
